changed create_pt_cloud to cluster only over xyz.  starting change to 3d feature library to have separate descriptor that holds auxiliary information
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@18725 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: febb099ad1a6a83706ad252ba1abdd2637e5a473

+++ sandbox/descriptors_3d/src/local_geometry.cpp
@@ -72,11 +72,11 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
       // Find neighboring points within radius
       neighbor_indices.clear();
       neighbor_distances.clear();
-      if (data_kdtree.radiusSearch(*(interest_pts[(size_t)i]), spectral_radius_, neighbor_indices, neighbor_distances)
+      if (data_kdtree.radiusSearch(*(interest_pts[i]), spectral_radius_, neighbor_indices, neighbor_distances)
           == false)
       {
         ROS_ERROR("LocalGeometry::compute radius search failed");
-        results[(size_t)i].clear();
+        results[i].clear();
         continue;
       }
 
@@ -85,7 +85,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
       if (neighbor_indices.size() < 3)
       {
         ROS_ERROR("did not have enough neighbors");
-        results[(size_t)i].clear();
+        results[i].clear();
         continue;
       }
 
@@ -102,8 +102,8 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
 
       // ----------------------------------------
       // Compute features
-      populateFeaturesWithSpectral(eigen_values, tangent, normal, *(interest_pts[(size_t)i]), data, neighbor_indices,
-          results[(size_t)i]);
+      populateFeaturesWithSpectral(eigen_values, tangent, normal, *(interest_pts[i]), data, neighbor_indices,
+          results[i]);
     }
   }
   // Otherwise compute features that do not require normal/tangent information
@@ -111,7 +111,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
   {
     for (unsigned int i = 0 ; i < nbr_pts ; i++)
     {
-      populateFeatures(0, *(interest_pts[(size_t)i]), results[(size_t)i]);
+      populateFeatures(0, *(interest_pts[i]), results[i]);
     }
   }
 }
@@ -119,6 +119,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
 // --------------------------------------------------------------
 /*! See function definition */
 // --------------------------------------------------------------
+/*
 void LocalGeometry::compute(const robot_msgs::PointCloud& data,
                             cloud_kdtree::KdTree& data_kdtree,
                             cv::Vector<cv::Vector<float> >& results)
@@ -159,8 +160,8 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
       neighbor_distances.clear();
       if (data_kdtree.radiusSearch(i, spectral_radius_, neighbor_indices, neighbor_distances) == false)
       {
-        results[(size_t)i].clear();
-        ROS_ERROR("LocalGeometry::compute radius search failed, vector has size: %u", results[(size_t)i].size());
+        results[i].clear();
+        ROS_ERROR("LocalGeometry::compute radius search failed, vector has size: %u", results[i].size());
         continue;
       }
 
@@ -168,8 +169,8 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
       // Need 3-by-3 matrix to have full rank
       if (neighbor_indices.size() < 3)
       {
-        results[(size_t)i].clear();
-        ROS_ERROR("LocalGeometry::compute not enough neighbors at pt %u, vector has size: %u", i, results[(size_t)i].size());
+        results[i].clear();
+        ROS_ERROR("LocalGeometry::compute not enough neighbors at pt %u, vector has size: %u", i, results[i].size());
         continue;
       }
 
@@ -187,7 +188,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
       // ----------------------------------------
       // Compute features
       populateFeaturesWithSpectral(eigen_values, tangent, normal, data.pts[i], data, neighbor_indices,
-          results[(size_t)i]);
+          results[i]);
     }
   }
   // Otherwise compute features that do not require normal/tangent information
@@ -195,10 +196,11 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
   {
     for (unsigned int i = 0 ; i < nbr_pts ; i++)
     {
-      populateFeatures(0, data.pts[(size_t)i], results[(size_t)i]);
+      populateFeatures(0, data.pts[i], results[i]);
     }
   }
 }
+*/
 
 // --------------------------------------------------------------
 /*! See function definition */
@@ -230,7 +232,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
     {
       if (spectral_radius_ < 0.0)
       {
-        curr_region_indices = interest_region_indices[(size_t)i];
+        curr_region_indices = interest_region_indices[i];
       }
       else
       {
@@ -244,7 +246,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
             == false)
         {
           ROS_ERROR("LocalGeometry::compute radius search failed");
-          results[(size_t)i].clear();
+          results[i].clear();
           continue;
         }
       }
@@ -254,7 +256,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
       if (curr_region_indices->size() < 3)
       {
         ROS_ERROR("did not have enough neighbors");
-        results[(size_t)i].clear();
+        results[i].clear();
         continue;
       }
 
@@ -272,7 +274,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
       // ----------------------------------------
       // Compute features
       populateFeaturesWithSpectral(eigen_values, tangent, normal, centroid, data, *curr_region_indices,
-          results[(size_t)i]);
+          results[i]);
     }
 
   }
@@ -282,7 +284,7 @@ void LocalGeometry::compute(const robot_msgs::PointCloud& data,
     for (unsigned int i = 0 ; i < nbr_regions ; i++)
     {
       cloud_geometry::nearest::computeCentroid(data, *(curr_region_indices), centroid);
-      populateFeatures(0, centroid, results[(size_t)i]);
+      populateFeatures(0, centroid, results[i]);
     }
   }
 }
@@ -301,9 +303,9 @@ void LocalGeometry::populateFeaturesWithSpectral(const Eigen::Vector3d& eigen_va
   result.resize(result_size_);
 
   unsigned int idx = 0;
-  result[(size_t)idx++] = eigen_values[0]; // scatter
-  result[(size_t)idx++] = eigen_values[2] - eigen_values[1]; // linear
-  result[(size_t)idx++] = eigen_values[1] - eigen_values[0]; // surface
+  result[idx++] = eigen_values[0]; // scatter
+  result[idx++] = eigen_values[2] - eigen_values[1]; // linear
+  result[idx++] = eigen_values[1] - eigen_values[0]; // surface
 
   if (ref_tangent_defined_)
   {
@@ -312,7 +314,7 @@ void LocalGeometry::populateFeaturesWithSpectral(const Eigen::Vector3d& eigen_va
     {
       tangent_dot = tangent.dot(ref_tangent_flipped_);
     }
-    result[(size_t)idx++] = tangent_dot;
+    result[idx++] = tangent_dot;
   }
 
   if (ref_normal_defined_)
@@ -322,7 +324,7 @@ void LocalGeometry::populateFeaturesWithSpectral(const Eigen::Vector3d& eigen_va
     {
       normal_dot = normal.dot(ref_normal_flipped_);
     }
-    result[(size_t)idx++] = normal_dot;
+    result[idx++] = normal_dot;
   }
 
   populateFeatures(idx, interest_pt, result);
@@ -342,6 +344,6 @@ void LocalGeometry::populateFeatures(unsigned int start_idx,
 
   if (use_elevation_)
   {
-    result[(size_t)start_idx++] = interest_pt.z;
+    result[start_idx++] = interest_pt.z;
   }
 }
